Adds EventMachine timeout configuration
On my current Rails project, we're faced with a particularly slow asset
compilation step. This has led to us having flakey tests, as our Puffing
Billy tests using the selenium_billy driver often timeout when compiling
assets. It'd be nice to be able to change the configuration for the
EventMachine::HttpRequest that the proxy handler uses so that we can
safely deal with slow requests.