Don't recommend changing javascript_driver config
javascript_driver is a configuration value that changes the driver used for *all* scenarios tagged with @javascript after it is set. That means that it impacts tests that are run after a @billy test. For me, this manifested as a test suite that froze partway through and timed out. Our @billy tests should not impact other tests, and setting current_driver is sufficient to make sure this test uses billy, regardless of the presence of @javascript.