Support distinguishing cached non-POST requests based on body
This adds a config option `cache_request_body_methods`, allow you to
specify the HTTP methods for which you would like the cache to
distinguish between requests with different bodies. Previously, this
only applied to POST requests, but PATCH and PUT requests also often
have different bodies that it might be useful to distinguish between
while testing. The default value remains 'post' so the functionality
when the option is not set should not change.

I also added a before_install step to try to fix the gem install errors
I was seeing on Travis CI.