Catch unhandled request handler errors
This fixes an issue where when an underlying error happens, say someone
is using VCR for client side stubs and it raises an error, the web
process hangs due to the proxy crashing. This catches any unhandled
standard error and returns it as part of the error hash from the
request. This will result in the error message being written to the log
and allowing the proxy to continue running.