Rescue Errno::ECONNRESET in shutdown phase
While waiting for shutdown, the proxy attempts to connect to the
event machine TCP socket in a loop until the connection fails to be
made.

There's a small possibility here that the connection is established but
the socket is closed by the remote before the poller closes it, in which
case we get an Errno::ECONNRESET exception that isn't currently handled.

I think this error also indicates that we've shut down successfully, so
this adds it to the existing rescue.