Update http-parser to 0.8.0
Puffing billy locks http-parser down to version 0.6.0 (for any project using this gem), which was released in 2013. The http-parser gem isn't updated very often but the latest version (0.8.0) was released in 2021.

I've updated the gemspec to use the latest version so Puffing Billy users can move their own projects to a more recent dependency.